fix: Use :latin1 encoding for stdio
The TLDR is the way to read and write "raw bytes" from stdio is to set 
the encoding to :latin1 and use the &IO.binread/2 and &IO.binwrite/2
functions.

Longer:

I ran into some strange issue where the Neovim LSP client was sending a
message that had Content-Length: 53866, and I confirmed in Lua that that
was the correct length of the string.

But when Elixir was reading it from stdio, I had to have been counting
two chars as the same chars, because the string it would read in was
53864, which caused it to overeat by 2 chars, eating the next header.

This would cause a Jason.decode! error sometimes, and other times it
would just hang because it was waiting for 2 more chars to come in.

I had dealt with this before in a way, but I'm not sure why it started
breaking again.