feat!: be resilient to LSP protocol errors
Previously, if a user passed non compliant reply to a request, the
framework would raise and your server would crash.

Previously, if a client sent an invalid request or notification, the
framework would raise, not providing a good error for the user or the
LSP author.

Now, in those three cases, if the framework sees a non-compliant
response, it will respond to the client with a relevant error code and
message.

Note: in the case of notifications, we just log, as there is nothing to
respond to.